🔨 Refactor: 컴포넌트 하위 _model 디렉토리 제거
