--- src/app/lobby/_component/Canvas/CharacterArray.ts
@@ -1,6 +1,0 @@
-const characterImages: Record<"character1" | "character2", string> = {
-  character1: "/character/character1.png",
-  character2: "/character/character2.png",
-};
-
-export default characterImages;
