--- src/app/lobby/_component/SolvedUsers/SolvedUsersContent.style.ts
@@ -1,23 +1,0 @@
-const Style = {
-  container: `
-      bg-[#f3d8ae]
-      bg-[url('/images/quest-window-bg.png')]
-      bg-repeat
-      border-2
-      border-[#8b4513]
-      rounded
-      p-4
-      text-black
-      [font-family:MaplestoryOTFBold,sans-serif]
-    `,
-  heading: `
-      text-lg
-      font-bold
-      mb-2
-    `,
-  message: `
-      mb-1
-    `,
-};
-
-export default Style;
