--- src/app/lobby/_model/DailyProblem.ts
@@ -1,9 +1,0 @@
-export interface DailyProblemProps {
-  dailyProblem: {
-    id: number;
-    title: string;
-    link: string;
-  } | null;
-  isProblemSolved: boolean;
-  handleSolveDailyProblem: () => void;
-}
