--- src/app/lobby/_model/Npc.ts
@@ -1,12 +1,0 @@
-export interface NpcInfo {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  image: string;
-  modalTitle: string;
-  name: string;
-}
-export interface NpcListProps {
-  npcs: NpcInfo[];
-}
