--- src/app/lobby/_model/SolvedUser.ts
@@ -1,11 +1,0 @@
-export interface SolvedUser {
-  userId: string;
-  nickname: string;
-  solvedProblemId: number;
-  solvedDate: string;
-}
-
-export interface SolvedUsersProps {
-  dailySolvedUsers: SolvedUser[];
-  getTodayString: () => string;
-}
