--- src/app/lobby/_model/User.ts
@@ -1,7 +1,0 @@
-export interface User {
-  id: string;
-  x: number;
-  y: number;
-  characterType: "character1" | "character2";
-  nickname: string;
-}
