--- src/app/myroom/[id]/_model/Portfolio.ts
@@ -1,7 +1,0 @@
-export interface PortfolioModalProps {
-  open: boolean;
-  onClose: (open: boolean) => void;
-  portfolioFile: File | null;
-  setPortfolioFile: React.Dispatch<React.SetStateAction<File | null>>;
-  onSave: () => void;
-}
