--- src/app/myroom/[id]/_model/Resume.ts
@@ -1,7 +1,0 @@
-export interface ResumeModalProps {
-  open: boolean;
-  onClose: (open: boolean) => void;
-  resumeLink: string;
-  setResumeLink: React.Dispatch<React.SetStateAction<string>>;
-  onSave: () => void;
-}
