--- src/app/myroom/[id]/_model/TechStack.ts
@@ -1,8 +1,0 @@
-export interface TechStackModalProps {
-  open: boolean;
-  onClose: (open: boolean) => void;
-  techStackList: string[];
-  selectedTech: string;
-  setSelectedTech: React.Dispatch<React.SetStateAction<string>>;
-  onSave: () => void;
-}
