+++ src/app/questmap/_component/Npc/NpcModal.tsx
@@ -8,8 +8,8 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
+import { NpcModalProps } from "@/model/Npc";
 
-import { NpcModalProps } from "../../_model/Npc";
 import Style from "./NpcModal.style";
 
 export const NpcModal: React.FC<NpcModalProps> = ({
