--- src/app/questmap/_model/Npc.ts
@@ -1,19 +1,0 @@
-export interface NpcInfo {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  image: string;
-  modalTitle: string;
-  name: string;
-}
-export interface NpcListProps {
-  npcs: NpcInfo[];
-}
-
-export interface NpcModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  title: string;
-  children?: React.ReactNode;
-}
