+++ src/model/NoticeBoard.ts
