+++ src/model/NpcModal.ts
