+++ src/model/Qna.ts
